usage of promise fromcallback
